+++ src/Message/SyncCustomersMessageHandler.php
@@ -7,13 +7,12 @@ namespace Listrak\Message;
 use Listrak\Service\DataMappingService;
 use Listrak\Service\ListrakApiService;
 use Psr\Log\LoggerInterface;
+use Shopware\Core\Checkout\Customer\CustomerCollection;
 use Shopware\Core\Framework\DataAbstractionLayer\Dbal\Common\RepositoryIterator;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Sorting\FieldSorting;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
-use Symfony\Component\Messenger\MessageBusInterface;
 
 #[AsMessageHandler]
 final class SyncCustomersMessageHandler
@@ -24,9 +23,10 @@ final class SyncCustomersMessageHandler
 
     private LoggerInterface $logger;
 
+    /**
+     * @param EntityRepository<CustomerCollection> $customerRepository
+     */
     public function __construct(
-        private readonly MessageBusInterface $messageBus,
-        private readonly EventDispatcherInterface $eventDispatcher,
         private readonly EntityRepository $customerRepository,
         DataMappingService $dataMappingService,
         ListrakApiService $listrakApiService,
@@ -39,7 +39,7 @@ final class SyncCustomersMessageHandler
 
     public function __invoke(SyncCustomersMessage $message): void
     {
-        $this->logger->notice('Full Listrak customer import started.');
+        $this->logger->notice('Full Listrak customer sync started.');
         $context = $message->getContext();
         try {
             $criteria = new Criteria();
@@ -49,11 +49,9 @@ final class SyncCustomersMessageHandler
             while (($result = $iterator->fetch()) !== null) {
                 $customers = $result->getEntities();
                 $items = [];
-                if (!empty($customers)) {
-                    foreach ($customers as $customer) {
-                        $item = $this->dataMappingService->mapCustomerData($customer);
-                        $items[] = $item;
-                    }
+                foreach ($customers as $customer) {
+                    $item = $this->dataMappingService->mapCustomerData($customer);
+                    $items[] = $item;
                 }
                 if (!empty($items)) {
                     $this->listrakApiService->importCustomer($items, $context);